moving animation __init__ functions to a local utils module
